Adding MutationObserver support to dynamically add quotes
