+++ package.json
@@ -29,13 +29,13 @@
     "test": "npm run build && tap test/*.js"
   },
   "devDependencies": {
+    "@webcomponents/webcomponentsjs": "git://github.com/webcomponents/webcomponentsjs#v0.7.24",
     "babel-loader": "^7.1.2",
     "babel-minify-webpack-plugin": "^0.2.0",
     "babel-preset-es2015": "^6.24.1",
     "jsdom": "^7.0.0",
     "parse5": "^3.0.2",
     "tap": "^10.3.0",
-    "uglifyjs-webpack-plugin": "^0.4.6",
     "webpack": "^3.6.0"
   },
   "engines": {
