+++ test/smartquotes.js
@@ -121,13 +121,34 @@ test('smartquotes()', t => {
   t.end();
 });
 
-test('smartquotes exists in a browser', function(t) {
+test('smartquotes in a browser', function(t) {
 
   jsdom.env({
     file: './test/fixtures/basic.html',
-    scripts: '../../dist/smartquotes.js', // path relative to file
+    scripts: ['../../node_modules/webcomponents.js/MutationObserver.js', '../../dist/smartquotes.js'],
     onload: function (window) {
-      t.ok(window.smartquotes);
+
+      test('it exists as a global variable', t => {
+        t.ok(window.smartquotes);
+        t.end();
+      });
+
+      test('it can react to DOM mutations', t => {
+        window.smartquotes.listen();
+        var addedNode = window.document.createElement('div');
+        var changedNode = window.document.createElement('div');
+        addedNode.innerHTML = 'Adding "this" node.';
+        changedNode.innerHTML = 'No quotes.';
+        window.document.body.appendChild(addedNode);
+        window.document.body.appendChild(changedNode);
+        changedNode.innerHTML = "Changing \"this\" node.";
+        setTimeout(() => {
+          t.equal(addedNode.innerHTML, "Adding \u201cthis\u201d node.");
+          t.equal(changedNode.innerHTML, "Changing \u201cthis\u201d node.");
+          t.end();
+        });
+      });
+
       t.end();
     }
   });
